test(forces): 按数据流接缝重构测试套件——contract/physics/config/container 四目录
- models/ 解散：契约测试进 contract/、物理规律验证进 physics/；
  8 个混合文件拆分为单接缝文件（ecom_srp、indirect_term、point_mass、
  relativistic、shadow、third_body、thrust、low_thrust）
- container/ 瘦身为纯编排：kepler/hohmann/leo_j2 归 physics/、
  leo_config 归 config/
- 去冗余：srp_shadow 几何测试 3 个（与 shadow.py 逐字重复）、kepler
  闭合 1 个（与 point_mass 重复）、variable_mass events 1 个（与
  container 已有测试重复）删除
- 归位：relativistic 缓存测试 → integrators/bindings/test_ephem_cache.py
  （星历缓存接缝）；relativistic config round-trip 并入 config/test_force_config.py
- config/ 增 test_low_thrust_config_propagation.py（FiniteBurn round-trip
  传播验收，ADR 0004）；修复首次重构的 2 个测试函数名 bug
