+++ tests/numerical/forces/contract/test_point_mass_gravity.py
@@ -1,14 +1,12 @@
-"""PointMassGravity 测试。
+"""PointMassGravity 定义与序列化契约。
 
-Python 单点 ``compute_acceleration`` 已按 issue #378 删除，加速度由 Rust
-``propagate_compiled`` 承载；本文件验证 ``to_rust_spec`` 序列化与
-``ForceModel`` 端到端传播的物理行为。
+验证 ``to_rust_spec`` 序列化（显式 mu / system fallback / 无 system 报错）
+与 PhysicalModel 子类关系；物理行为见 ``physics/test_point_mass_gravity.py``。
 """
 
-import numpy as np
 import pytest
 
-from e2m2e.algorithm.forces import ForceModel, PointMassGravity
+from e2m2e.algorithm.forces import PointMassGravity
 
 pytestmark = pytest.mark.force
 
@@ -53,24 +51,6 @@ def test_point_mass_gravity_no_system_no_mu_raises():
     force = PointMassGravity(body="EARTH")
     with pytest.raises(ValueError, match="gravitational_parameter"):
         force.to_rust_spec(None)
-
-
-def test_point_mass_gravity_rust_propagation_matches_two_body_solution():
-    """Rust compiled 传播点质量圆轨道，一个周期后回到初值。"""
-    system = _FakeSystem()
-    force = PointMassGravity(body="EARTH", mu=398600.4415)
-    fm = ForceModel(system, forces=[force])
-
-    r = 6778.0
-    v = np.sqrt(398600.4415 / r)
-    period = 2.0 * np.pi * np.sqrt(r**3 / 398600.4415)
-    y0 = np.array([r, 0.0, 0.0, 0.0, v, 0.0])
-
-    result = fm.propagate(y0, (0.0, period), t_eval=np.array([0.0, period]))
-
-    final = result["states"][-1]
-    np.testing.assert_allclose(final[:3], y0[:3], rtol=1e-6, atol=1e-8)
-    np.testing.assert_allclose(final[3:], y0[3:], rtol=1e-6, atol=1e-8)
 
 
 def test_point_mass_gravity_is_physical_model():
