+++ tests/numerical/forces/contract/test_thrust_direction_frame.py
