--- tests/numerical/forces/models/test_relativistic_correction.py
@@ -1,200 +1,0 @@
-"""相对论修正力模型测试。
-
-Python 单点 ``compute_acceleration`` 已按 issue #378 删除；相对论物理行为由
-Rust ``propagate_compiled`` 承载。本文件保留配置 round-trip 与 Rust 端到端
-传播验证。
-"""
-
-import numpy as np
-import pytest
-
-from e2m2e.algorithm.forces import ForceModel, GravityField, RelativisticCorrection
-from tests.numerical.forces.conftest import EARTH_RE, keplerian_to_cartesian
-
-pytestmark = pytest.mark.force
-
-
-# 地球角动量矢量近似值 (km²/s)，与 GMAT 自动计算结果同量级。
-_EARTH_ANGULAR_MOMENTUM = np.array([0.0, 0.0, 1.18e3])
-
-
-def test_config_round_trip():
-    """RelativisticCorrection 支持 ForceModel 配置往返。"""
-    from e2m2e.algorithm.forces.force_config import build_force, serialize_force
-
-    original = RelativisticCorrection(
-        central_body="Earth",
-        primary_body="Sun",
-        enable_schwarzschild=True,
-        enable_lense_thirring=False,
-        enable_de_sitter=True,
-        angular_momentum_vector=[0.0, 0.0, 7.5e33],
-        body_radius=6378.137,
-        c=299792.458,
-        gamma=1.0,
-    )
-
-    config = serialize_force(original)
-    restored = build_force(config["type"], config["params"])
-
-    assert isinstance(restored, RelativisticCorrection)
-    assert restored.central_body == "EARTH"
-    assert restored.primary_body == "SUN"
-    assert restored.enable_schwarzschild is True
-    assert restored.enable_lense_thirring is False
-    assert restored.enable_de_sitter is True
-    np.testing.assert_array_equal(restored.angular_momentum_vector, np.array([0.0, 0.0, 7.5e33]))
-    assert restored.body_radius == pytest.approx(6378.137)
-    assert restored.c == pytest.approx(299792.458)
-    assert restored.gamma == pytest.approx(1.0)
-
-
-@pytest.mark.spice
-def test_gps_relativistic_position_difference_magnitude(earth_icrf_system):
-    """GPS 轨道 1 天传播，相对论修正导致可观测的终端位置漂移。"""
-    system = earth_icrf_system
-    mu = system.gravitational_parameter("EARTH")
-
-    # GPS 类轨道
-    a0 = 26560.0
-    y0 = keplerian_to_cartesian(a0, 0.0, 55.0, 0.0, 0.0, 0.0, mu)
-
-    et0 = system.spice.utc_to_et("2025-06-21T11:00:06")
-    t_span = (et0, et0 + 86400.0)
-    t_eval = np.array([et0, et0 + 86400.0])
-
-    gravity = GravityField(body="EARTH", degree=2, order=0)
-    fm_without = ForceModel(system, forces=[gravity])
-    result_without = fm_without.propagate(y0, t_span, t_eval=t_eval, max_steps=200_000)
-
-    relcorr = RelativisticCorrection(
-        central_body="EARTH",
-        angular_momentum_vector=_EARTH_ANGULAR_MOMENTUM,
-    )
-    fm_with = ForceModel(system, forces=[gravity, relcorr])
-    result_with = fm_with.propagate(y0, t_span, t_eval=t_eval, max_steps=200_000)
-
-    pos_diff = np.linalg.norm(result_with["states"][-1, :3] - result_without["states"][-1, :3])
-    # 物理量级：GPS 轨道 Schwarzschild 修正约 0.3 mm/天（3e-7 km）。
-    # 下界收到 1/10 物理量级以防回归把数量级改坏（远低于物理 3 个数量级 → 收紧到 0.1×）。
-    # 上界 0.01 km（10 cm/天）覆盖 Lense-Thirring / de Sitter 等次级项贡献。
-    assert 3e-8 <= pos_diff <= 0.01, f"GPS 1-day position diff = {pos_diff:.6e} km"
-
-
-@pytest.mark.spice
-def test_leo_relativistic_position_difference_magnitude(earth_icrf_system):
-    """LEO 轨道 1 天传播，相对论修正导致可观测的终端位置漂移。"""
-    system = earth_icrf_system
-    mu = system.gravitational_parameter("EARTH")
-
-    r_earth = EARTH_RE
-    a0 = r_earth + 400.0
-    y0 = keplerian_to_cartesian(a0, 0.0, 51.6, 0.0, 0.0, 0.0, mu)
-
-    et0 = system.spice.utc_to_et("2025-06-21T11:00:06")
-    t_span = (et0, et0 + 86400.0)
-    t_eval = np.array([et0, et0 + 86400.0])
-
-    gravity = GravityField(body="EARTH", degree=2, order=0)
-    fm_without = ForceModel(system, forces=[gravity])
-    result_without = fm_without.propagate(y0, t_span, t_eval=t_eval, max_steps=200_000)
-
-    relcorr = RelativisticCorrection(
-        central_body="EARTH",
-        angular_momentum_vector=_EARTH_ANGULAR_MOMENTUM,
-    )
-    fm_with = ForceModel(system, forces=[gravity, relcorr])
-    result_with = fm_with.propagate(y0, t_span, t_eval=t_eval, max_steps=200_000)
-
-    pos_diff = np.linalg.norm(result_with["states"][-1, :3] - result_without["states"][-1, :3])
-    # 物理量级：LEO Schwarzschild 修正约 2.5 mm/天（2.5e-6 km）。
-    # 下界收到 1/10 物理量级以防回归把数量级改坏（原 1e-3 km 比物理宽 2.7 个数量级 → 收紧到 0.1×）。
-    # 上界 0.01 km（10 cm/天）覆盖 Lense-Thirring / de Sitter 等次级项贡献。
-    assert 2.5e-7 <= pos_diff <= 0.01, f"LEO 1-day position diff = {pos_diff:.6e} km"
-
-
-@pytest.mark.spice
-def test_relativistic_cache_zero_ffi_and_consistency(earth_icrf_system):
-    """#268 验收 1+2（缓存路径）：relativity=1 传播全程零 cspice FFI，且与无缓存一致。
-
-    - 启用星历缓存（含 de Sitter 的 EARTH/SUN 相对 SSB + LT 的 ITRF93→J2000 sxform）
-    - ``propagate_compiled`` 走纯 Rust 相对论力，``ephem_ffi_call_count()`` 应为 0
-    - 缓存 vs 无缓存末态一致（< 1e-5 km）
-    """
-    from e2m2e.integrators import (
-        RkMethod,
-        disable_ephem_cache,
-        ephem_ffi_call_count,
-        propagate_compiled,
-        reset_ephem_ffi_call_count,
-    )
-
-    system = earth_icrf_system
-    mu_earth = system.gravitational_parameter("EARTH")
-    mu_sun = system.gravitational_parameter("SUN")
-
-    # LEO 轨道
-    r_earth = EARTH_RE
-    a0 = r_earth + 400.0
-    y0 = keplerian_to_cartesian(a0, 0.0, 51.6, 0.0, 0.0, 0.0, mu_earth)
-
-    et0 = system.spice.utc_to_et("2025-06-21T11:00:06")
-    t_eval = np.array([et0, et0 + 3600.0])
-
-    # 相对论力：Schwarzschild + LT(自动角动量→sxform) + de Sitter(→SSB 状态)
-    rel_spec = (
-        "relativistic",
-        "EARTH",
-        "SUN",
-        mu_earth,
-        mu_sun,
-        True,
-        True,
-        True,
-        None,  # angular_momentum_vector=None → 每步自动 sxform
-        None,  # body_radius → 默认表
-        1.0,
-    )
-
-    def run():
-        return propagate_compiled(
-            RkMethod.PD45,
-            float(et0),
-            [float(x) for x in y0],
-            600.0,
-            1e-10,
-            [float(x) for x in t_eval],
-            "EARTH",
-            [rel_spec],
-            200_000,
-        )
-
-    # 基线（无缓存）
-    res_base = run()
-
-    # 启用缓存：manager 会注册 EARTH/SUN 相对 SSB（de Sitter 用）+ name/ID 双键；
-    # sxform_pairs 注册 ITRF93→J2000 6×6 变换（LT 自动角动量用）。
-    system.spice.enable_ephem_cache(
-        ["EARTH", "SUN"],
-        et0,
-        et0 + 3600.0,
-        dt=600.0,
-        observer="EARTH",
-        frame_pairs=[("ITRF93", "J2000")],
-        sxform_pairs=[("ITRF93", "J2000")],
-    )
-    try:
-        reset_ephem_ffi_call_count()
-        res_cached = run()
-        ffi_during_prop = ephem_ffi_call_count()
-        # #268 验收标准 1：relativity=1 下传播全程零 cspice FFI
-        assert ffi_during_prop == 0, (
-            f"传播期间 cspice FFI 调用 {ffi_during_prop} 次，应为 0（LT sxform / de Sitter "
-            "spkezr 均应走缓存）"
-        )
-    finally:
-        disable_ephem_cache()
-
-    # 缓存 vs 无缓存一致
-    diff = np.linalg.norm(np.asarray(res_cached["states"][-1]) - np.asarray(res_base["states"][-1]))
-    assert diff < 1e-5, f"缓存 vs 无缓存末态差异 {diff:.3e} km 超过 1e-5"
