+++ tests/numerical/integrators/bindings/test_ephem_cache.py
@@ -2,9 +2,10 @@
 
 验证 ``e2m2e.integrators`` 的 ``enable_ephem_cache`` / ``disable_ephem_cache``
 绑定 API：
-- 激活缓存 vs 未激活，传播末态数值一致（< 1e-6 km）
+- 激活缓存 vs 未激活，传播末态数值一致
 - 未激活时零回归（与现有行为逐字一致）
-- 缓存对含第三体的传播同样有效
+- 缓存对含第三体、固体潮、相对论力的传播同样有效
+- 相对论力传播全程零 cspice FFI（#268）
 
 归位说明：本文件验证的是星历缓存基础设施（pyo3 绑定层），力模型只是
 传播载体，故位于 integrators/bindings/ 而非 forces/。
@@ -40,7 +41,7 @@ def earth_system(spice_kernel_path):
     spice.load_kernel(spice_kernel_path)
     bf = load_body_fixed_kernels(spice)
     try:
-        system = EphemerisSystem(bodies=["EARTH"], spice=spice, origin="EARTH")
+        system = EphemerisSystem(bodies=["EARTH", "SUN"], spice=spice, origin="EARTH")
         system.coordinate_system = CoordinateSystem(
             axes=ICRSAxes(),
             origin=CelestialBodyOrigin(body="EARTH", spice=spice),
@@ -184,3 +185,89 @@ def test_ephem_cache_tide_solid_consistency(earth_system):
 
     diff = np.linalg.norm(state_cached - state_baseline)
     assert diff < 1e-2, f"tide=1 缓存 vs 未缓存末态差异 {diff:.3e} km 超过 1e-2"
+
+
+@pytest.mark.spice
+def test_ephem_cache_relativistic_zero_ffi_and_consistency(earth_system):
+    """#268 验收 1+2（缓存路径）：relativity=1 传播全程零 cspice FFI，且与无缓存一致。
+
+    - 启用星历缓存（含 de Sitter 的 EARTH/SUN 相对 SSB + LT 的 ITRF93→J2000 sxform）
+    - ``propagate_compiled`` 走纯 Rust 相对论力，``ephem_ffi_call_count()`` 应为 0
+    - 缓存 vs 无缓存末态一致（< 1e-5 km）
+    """
+    from e2m2e.integrators import (
+        RkMethod,
+        ephem_ffi_call_count,
+        propagate_compiled,
+        reset_ephem_ffi_call_count,
+    )
+
+    system = earth_system
+    mu_earth = system.gravitational_parameter("EARTH")
+    mu_sun = system.gravitational_parameter("SUN")
+
+    # LEO 圆轨道
+    r0 = 6378.137 + 400.0
+    v0 = np.sqrt(mu_earth / r0)
+    y0 = np.array([r0, 0.0, 0.0, 0.0, v0, 0.0])
+
+    et0 = system.spice.utc_to_et("2025-06-21T11:00:06")
+    t_eval = np.array([et0, et0 + 3600.0])
+
+    # 相对论力：Schwarzschild + LT(自动角动量→sxform) + de Sitter(→SSB 状态)
+    rel_spec = (
+        "relativistic",
+        "EARTH",
+        "SUN",
+        mu_earth,
+        mu_sun,
+        True,
+        True,
+        True,
+        None,  # angular_momentum_vector=None → 每步自动 sxform
+        None,  # body_radius → 默认表
+        1.0,
+    )
+
+    def run():
+        return propagate_compiled(
+            RkMethod.PD45,
+            float(et0),
+            [float(x) for x in y0],
+            600.0,
+            1e-10,
+            [float(x) for x in t_eval],
+            "EARTH",
+            [rel_spec],
+            200_000,
+        )
+
+    # 基线（无缓存）
+    res_base = run()
+
+    # 启用缓存：manager 会注册 EARTH/SUN 相对 SSB（de Sitter 用）+ name/ID 双键；
+    # sxform_pairs 注册 ITRF93→J2000 6×6 变换（LT 自动角动量用）。
+    system.spice.enable_ephem_cache(
+        ["EARTH", "SUN"],
+        et0,
+        et0 + 3600.0,
+        dt=600.0,
+        observer="EARTH",
+        frame_pairs=[("ITRF93", "J2000")],
+        sxform_pairs=[("ITRF93", "J2000")],
+    )
+    try:
+        reset_ephem_ffi_call_count()
+        res_cached = run()
+        ffi_during_prop = ephem_ffi_call_count()
+        # #268 验收标准 1：relativity=1 下传播全程零 cspice FFI
+        assert ffi_during_prop == 0, (
+            f"传播期间 cspice FFI 调用 {ffi_during_prop} 次，应为 0（LT sxform / de Sitter "
+            "spkezr 均应走缓存）"
+        )
+    finally:
+        system.spice.disable_ephem_cache()
+
+    # 缓存 vs 无缓存一致
+    diff = np.linalg.norm(np.asarray(res_cached["states"][-1]) - np.asarray(res_base["states"][-1]))
+    assert diff < 1e-5, f"缓存 vs 无缓存末态差异 {diff:.3e} km 超过 1e-5"
